third_party/WebKit/LayoutTests/fast/css/last-child-innerhtml-expected.txt

Check if replacing innerHTML handles :last-child properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Replacing three children.
PASS getComputedStyle(element).getPropertyValue('color') is 'rgb(0, 0, 255)'
PASS getComputedStyle(element).getPropertyValue('color') is 'rgb(0, 0, 0)'
PASS getComputedStyle(element).getPropertyValue('color') is 'rgb(128, 0, 128)'
Replacing one child.
PASS getComputedStyle(element).getPropertyValue('color') is 'rgb(0, 0, 255)'
PASS getComputedStyle(element).getPropertyValue('color') is 'rgb(0, 0, 0)'
PASS getComputedStyle(element).getPropertyValue('color') is 'rgb(128, 0, 128)'
PASS successfullyParsed is true

TEST COMPLETE
